[CAS-391] Move InternalStreamChatApi to core module, mark FSM as internal
