+++ stream-chat-android-client/build.gradle
@@ -60,7 +60,7 @@ tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
                 '-progressive',
                 '-Xexplicit-api=strict',
                 '-Xopt-in=kotlin.RequiresOptIn',
-                '-Xopt-in=io.getstream.chat.android.client.internal.InternalStreamChatApi',
+                '-Xopt-in=io.getstream.chat.core.internal.InternalStreamChatApi',
         ]
     }
 }
