+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/internal/DispatcherProvider.kt
@@ -2,6 +2,7 @@ package io.getstream.chat.android.client.internal
 
 import io.getstream.chat.android.client.internal.DispatcherProvider.reset
 import io.getstream.chat.android.client.internal.DispatcherProvider.set
+import io.getstream.chat.core.internal.InternalStreamChatApi
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
