+++ stream-chat-android-ui-components/build.gradle
@@ -65,7 +65,7 @@ tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
                 '-progressive',
                 '-Xexplicit-api=strict',
                 '-Xopt-in=kotlin.contracts.ExperimentalContracts',
-                '-Xopt-in=io.getstream.chat.android.client.internal.InternalStreamChatApi'
+                '-Xopt-in=io.getstream.chat.core.internal.InternalStreamChatApi'
         ]
     }
 }
