+++ stream-chat-android/build.gradle
@@ -73,12 +73,13 @@ tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
         freeCompilerArgs += [
                 '-progressive',
                 '-Xexplicit-api=strict',
-                '-Xopt-in=io.getstream.chat.android.client.internal.InternalStreamChatApi'
+                '-Xopt-in=io.getstream.chat.core.internal.InternalStreamChatApi'
         ]
     }
 }
 
 dependencies {
+    implementation project(":stream-chat-core")
     api project(":stream-chat-android-offline")
 
     implementation Dependencies.kotlinStdLib
