+++ stream-chat-android/src/main/java/com/getstream/sdk/chat/ImageLoader.kt
@@ -18,7 +18,7 @@ import coil.transform.CircleCropTransformation
 import coil.transform.GrayscaleTransformation
 import coil.transform.RoundedCornersTransformation
 import io.getstream.chat.android.client.internal.DispatcherProvider
-import io.getstream.chat.android.client.internal.InternalStreamChatApi
+import io.getstream.chat.core.internal.InternalStreamChatApi
 import kotlinx.coroutines.withContext
 import coil.loadAny as coilLoadAny
