+++ stream-chat-core/build.gradle
@@ -14,7 +14,9 @@ tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
     kotlinOptions {
         freeCompilerArgs += [
                 '-progressive',
-                '-Xexplicit-api=strict'
+                '-Xexplicit-api=strict',
+                '-Xopt-in=kotlin.RequiresOptIn',
+                '-Xopt-in=io.getstream.chat.core.internal.InternalStreamChatApi',
         ]
     }
 }
