--- stream-chat-core/src/main/java/io/getstream/chat/core/functional/fsm/builder/FSMBuilderMarker.kt
@@ -1,4 +1,0 @@
-package io.getstream.chat.core.functional.fsm.builder
-
-@DslMarker
-public annotation class FSMBuilderMarker
