+++ stream-chat-core/src/main/java/io/getstream/chat/core/internal/fsm/Event.kt
@@ -1,0 +1,6 @@
+package io.getstream.chat.core.internal.fsm
+
+import io.getstream.chat.core.internal.InternalStreamChatApi
+
+@InternalStreamChatApi
+public interface Event
