+++ stream-chat-core/src/main/java/io/getstream/chat/core/internal/fsm/builder/FSMBuilder.kt
@@ -1,11 +1,13 @@
-package io.getstream.chat.core.functional.fsm.builder
+package io.getstream.chat.core.internal.fsm.builder
 
-import io.getstream.chat.core.functional.fsm.Event
-import io.getstream.chat.core.functional.fsm.FiniteStateMachine
-import io.getstream.chat.core.functional.fsm.State
-import io.getstream.chat.core.functional.fsm.StateFunction
+import io.getstream.chat.core.internal.InternalStreamChatApi
+import io.getstream.chat.core.internal.fsm.Event
+import io.getstream.chat.core.internal.fsm.FiniteStateMachine
+import io.getstream.chat.core.internal.fsm.State
+import io.getstream.chat.core.internal.fsm.StateFunction
 import kotlin.reflect.KClass
 
+@InternalStreamChatApi
 @FSMBuilderMarker
 public class FSMBuilder<S : State, E : Event> {
     private lateinit var _initialState: S
